drop safe client completions
client completions were not sending cancellation messages when the
completion was dropped. The rpc model has cancellation, and the server
has tests to ensure it considers them, but the rpc client was not
handled.

Now when a streaming or unary rpc is dropped without being gracefully
completed, a cancellation control code will be sent.