Remove Send bound
This required moving the connection customization into the SocketListener
trait. It kind of makes sense there anyway. But in this way, there
is enough flexibility that none of the internal `protosocket-server` or
`protosocket-rpc::server` infrastructure should depend on `Send`.

Of course, it's easiest to just use the Tokio spawn with Send and all,
but if you need to do something else, this gets you closer.